requirements.txt
mysql-connector-python
Pillow
matplotlib
reportlab
tkcalendar